Move Keras model building into model_examples.
PiperOrigin-RevId: 232967013
